Kubernetes runtime files (#36)
* Move prepare_to_upload to infractl.fs module

* Make isort, black, pylint happy

* Runtime files for Kubernetes
